# web/openapi/faultevolve.openapi.json
{
  "components": {
    "schemas": {
...
      "ArtifactState": {
        "description": "Availability of one artifact inside a run directory.",
        "enum": [
          "available",
          "missing",
          "invalid"
        ],
        "title": "ArtifactState",
        "type": "string"
      },
...
      "DiscoveryResponse": {
        "additionalProperties": false,
        "description": "Artifacts of the knowledge-discovery pipeline (TODO §6.2 / §6.3).\n\n``phenomena`` / ``preregistration`` / ``controls`` mirror what\n``discovery/pipeline.py`` writes: all three are JSON *lists* (one entry per\nclaim / preregistration / controls snapshot), not objects.",
        "properties": {
          "certificates": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Certificates",
            "type": "array"
          },
          "claims": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Claims",
            "type": "array"
          },
          "claims_state": {
            "$ref": "#/components/schemas/ArtifactState",
            "default": "missing"
          },
          "contract_version": {
            "title": "Contract Version",
            "type": "string"
          },
          "controls": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Controls",
            "type": "array"
          },
          "controls_state": {
            "$ref": "#/components/schemas/ArtifactState",
            "default": "missing"
          },
          "discovered_card_ids": {
            "items": {
              "type": "string"
            },
            "title": "Discovered Card Ids",
            "type": "array"
          },
          "discovered_cards_state": {
            "$ref": "#/components/schemas/ArtifactState",
            "default": "missing"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Enabled"
          },
          "matches": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Matches",
            "type": "array"
          },
          "matches_state": {
            "$ref": "#/components/schemas/ArtifactState",
            "default": "missing"
          },
          "mechanisms_state": {
            "$ref": "#/components/schemas/ArtifactState",
            "default": "missing"
          },
          "phenomena": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Phenomena",
            "type": "array"
          },
          "phenomena_state": {
            "$ref": "#/components/schemas/ArtifactState",
            "default": "missing"
          },
          "preregistration": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Preregistration",
            "type": "array"
          },
          "preregistration_state": {
            "$ref": "#/components/schemas/ArtifactState",
            "default": "missing"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "theories": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Theories",
            "type": "array"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/ContractWarning"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "contract_version",
          "run_id"
        ],
        "title": "DiscoveryResponse",
        "type": "object"
      },
      "EventPageResponse": {
        "additionalProperties": false,
        "description": "Incremental event page for polling / SSE catch-up.",
        "properties": {
          "contract_version": {
            "title": "Contract Version",
            "type": "string"
          },
          "data_state": {
            "$ref": "#/components/schemas/RunDataState"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/EventResponse"
            },
            "title": "Events",
            "type": "array"
          },
          "has_more": {
            "default": false,
            "title": "Has More",
            "type": "boolean"
          },
          "last_event_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Last Event Id"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/ContractWarning"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "contract_version",
          "run_id",
          "events",
          "data_state"
        ],
        "title": "EventPageResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InsightsResponse": {
        "additionalProperties": false,
        "description": "Insights for a run.\n\n``available`` distinguishes the two empty cases the UI must not conflate:\nthe run never carried an ``insight`` table at all (``available=false``) vs.\nthe table exists and simply holds no rows (``available=true`` + ``[]``).",
        "properties": {
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "contract_version": {
            "title": "Contract Version",
            "type": "string"
          },
          "insights": {
            "items": {
              "$ref": "#/components/schemas/InsightResponse"
            },
            "title": "Insights",
            "type": "array"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/ContractWarning"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "contract_version",
          "run_id",
          "available"
        ],
        "title": "InsightsResponse",
        "type": "object"
      },
      "KnowledgeCardResponse": {
        "additionalProperties": false,
        "description": "One knowledge card plus how the run actually used it.\n\n``local_label`` is a UI-only marker and must be displayed together with the\nnote that it does not affect the run.\n\nProse (``title``, ``category``, ``claim``, ``source_kind``, ``conditions``,\n``priority``, ``sources``) is copied from the task knowledge catalog when\nthe run registry names that task and the card id matches. A run directory\nstill does not store the prose. Unmatched cards keep these fields empty\nrather than receiving invented text. ``priority`` is the catalog's 1–5\nrank, not a value derived from ``mean_delta``.",
        "properties": {
          "adopted_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Adopted Count"
          },
          "adopted_iterations": {
            "items": {
              "type": "integer"
            },
            "title": "Adopted Iterations",
            "type": "array"
          },
          "adopted_node_ids": {
            "items": {
              "type": "string"
            },
            "title": "Adopted Node Ids",
            "type": "array"
          },
          "affects_run": {
            "default": false,
            "title": "Affects Run",
            "type": "boolean"
          },
          "card_id": {
            "title": "Card Id",
            "type": "string"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Category"
          },
          "claim": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Claim"
          },
          "conditions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Conditions"
          },
          "invalid_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Invalid Count"
          },
          "local_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Local Label"
          },
          "mean_delta": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Mean Delta"
          },
          "offered_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Offered Count"
          },
          "priority": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Priority"
          },
          "refuted_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Refuted Count"
          },
          "source_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Source Detail"
          },
          "source_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Source Kind"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeSourceResponse"
            },
            "title": "Sources",
            "type": "array"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Title"
          }
        },
        "required": [
          "card_id"
        ],
        "title": "KnowledgeCardResponse",
        "type": "object"
      },
...
      "PresetFieldSource": {
        "additionalProperties": false,
        "description": "Where one field of a preset came from (TODO 4.6).\n\nLets the UI explain \"this number is the preset default\" versus \"this came\nfrom your patch\", without the front-end re-deriving the merge.",
        "properties": {
          "field": {
            "title": "Field",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "value": {
            "default": null,
            "title": "Value"
          }
        },
        "required": [
          "field",
          "source"
        ],
        "title": "PresetFieldSource",
        "type": "object"
      },
...
      "PresetResponse": {
        "additionalProperties": false,
        "description": "A complete, already-validated ``EvolveConfig`` for one preset.\n\n``config`` is the full config dump, not a patch: the front-end sends back\nonly a patch over this, and the server merges and re-validates.",
        "properties": {
          "config": {
            "additionalProperties": true,
            "title": "Config",
            "type": "object"
          },
          "display_name_zh": {
            "title": "Display Name Zh",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/PresetFieldSource"
            },
            "title": "Sources",
            "type": "array"
          },
          "summary_zh": {
            "default": "",
            "title": "Summary Zh",
            "type": "string"
          }
        },
        "required": [
          "name",
          "display_name_zh"
        ],
        "title": "PresetResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Read-only control and replay surface for the FaultEvolve engine. Scores are produced by the task evaluator and only ever read back here.",
    "title": "FaultEvolve Web API",
    "version": "1.11.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/presets/{name}": {
      "get": {
        "operationId": "get_preset_api_presets__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "读取单档预设",
        "tags": [
          "presets"
        ]
      }
    },
    "/api/presets/{name}/merge": {
      "post": {
        "description": "Merge, re-validate, and report per-field provenance.\n\nThe body is the raw ``config_patch`` object (``{\"budget\":\n{\"max_iterations\": 3}}``). Phase 7's run-creation endpoint will send\nthe same patch inside ``RunCreateRequest``; validating the merge\nsemantics here means that endpoint cannot discover them late.",
        "operationId": "merge_preset_api_presets__name__merge_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "把用户 patch 合并到预设并重新校验",
        "tags": [
          "presets"
        ]
      }
    },
...
    "/api/runs/{run_id}/discovery": {
      "get": {
        "description": "Discovery artifacts, each with its own availability state.",
        "operationId": "get_discovery_api_runs__run_id__discovery_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoveryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "知识发现产物",
        "tags": [
          "knowledge"
        ]
      }
    },
    "/api/runs/{run_id}/events": {
      "get": {
        "description": "One page of ``events.jsonl``, keyed on the engine's own event ids.\n\n``after_id`` is what makes reconnection work: the client sends the last\nid it saw and receives only what came after (PRD §16.3).",
        "operationId": "get_events_api_runs__run_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "After Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 2000,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "按事件序号增量拉取事件",
        "tags": [
          "events"
        ]
      }
    },
    "/api/runs/{run_id}/insights": {
      "get": {
        "description": "Insights, with an explicit ``available`` flag for the empty case.",
        "operationId": "get_insights_api_runs__run_id__insights_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsightsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "反思洞见",
        "tags": [
          "knowledge"
        ]
      }
    },
    "/api/runs/{run_id}/knowledge/cards": {
      "get": {
        "description": "One entry per card the run offered, with how it fared.",
        "operationId": "get_knowledge_cards_api_runs__run_id__knowledge_cards_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/KnowledgeCardResponse"
                  },
                  "title": "Response Get Knowledge Cards Api Runs  Run Id  Knowledge Cards Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "知识卡与运行中的采用统计",
        "tags": [
          "knowledge"
        ]
      }
    },
    "/api/runs/{run_id}/leaderboard": {
      "get": {
        "description": "Nodes ranked by score, with unscored nodes sorted last.",
        "operationId": "get_leaderboard_api_runs__run_id__leaderboard_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScoreboardEntryResponse"
                  },
                  "title": "Response Get Leaderboard Api Runs  Run Id  Leaderboard Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "排行榜",
        "tags": [
          "events"
        ]
      }
    },
...
    "/api/runs/{run_id}/scores": {
      "get": {
        "description": "One point per evaluated node, straight from the evaluator's record.",
        "operationId": "get_scores_api_runs__run_id__scores_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScorePointResponse"
                  },
                  "title": "Response Get Scores Api Runs  Run Id  Scores Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "分数时间线",
        "tags": [
          "events"
        ]
      }
    },
    "/api/runs/{run_id}/stream": {
      "get": {
        "description": "Server-sent events for one run, then a finite close.\n\nThe cursor comes from ``?after_id=`` or from the ``Last-Event-ID`` header\na reconnecting browser sends, and **the header wins**.\n\nThat precedence is the whole reason the header exists. ``EventSource``\nreconnects to the *same URL*, so a stream opened as\n``/stream?after_id=40`` comes back as ``?after_id=40`` **plus**\n``Last-Event-ID: 57`` -- and 57 is the truth, because that is what the\nclient actually received. Preferring the query string would re-deliver\neverything after 40 on every automatic reconnect, which PRD §16.3 names\ndirectly (\"同一事件重复投递不会产生重复节点或重复轮次\").\n\nBoth go through the same ``coerce_last_event_id``, so the two paths\ncannot come to mean different things.\n\nNo ``response_model`` on purpose: the body is a stream of frames, not a\ndocument, and declaring a model would make FastAPI try to validate -- and\nbuffer -- an endless body.\n\nThe stream is read-only, so like ``/live`` it needs no control flag.",
        "operationId": "stream_events_api_runs__run_id__stream_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "After Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "SSE 实时事件流（阶段 8）",
        "tags": [
          "events"
        ]
      }
    },
...
  }
}